remove dep & improve perf

+++ Auios.QuadTree/IQuadTreeObjectBounds.cs
@@ -5,9 +5,9 @@ namespace Auios.QuadTree
 {
     public interface IQuadTreeObjectBounds<in T>
     {
-        float GetTop(T obj);
-        float GetBottom(T obj);
-        float GetLeft(T obj);
-        float GetRight(T obj);
+        double GetTop(T obj);
+        double GetBottom(T obj);
+        double GetLeft(T obj);
+        double GetRight(T obj);
     }
 }
